feat: running working from the ui and showing results
